[UI] #81 KeyboardPresentationReadable 파일 별도 분리 (리뷰 반영)
